1st Submit to Testflight
Added icons and touched up some stuff. I got the notes to update to the questions and the bug fixed that shrunk the cells notes label when you would click another cell from that expanded cell. I also added a circle button in the bottom right of the sessionViewController to add the question and then made the button up top for info on the session.